+++ lib/recipes/information-systems/recipe.rb
@@ -9,5 +9,30 @@ do |doc, _resources|
 
   book.search('cnx-pi').trash
 
+  BakePreface.v1(book: book)
+
+  BakeUnitTitle.v1(book: book)
+  BakeChapterTitle.v1(book: book)
+  BakeChapterIntroductions.v1(book: book)
+
+  book.chapters.each do |chapter|
+    BakeNonIntroductionPages.v1(chapter: chapter)
+  end
+
+  # Appendix
+  book.pages('$.appendix').each do |page|
+    appendix_letter = [*('A'..'Z')][page.count_in(:book) - 1]
+
+    page.figures(only: :figure_to_number?).each do |figure|
+      BakeFigure.v1(figure: figure, number: "#{appendix_letter}#{figure.count_in(:page)}")
+    end
+
+    page.tables('$:not(.unnumbered)').each do |table|
+      BakeNumberedTable.v1(table: table, number: "#{appendix_letter}#{table.count_in(:page)}")
+    end
+    BakeAppendix.v1(page: page, number: appendix_letter)
+  end
+
+  BakeToc.v1(book: book)
   BakeFolio.v1(book: book)
 end